requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pillow==10.1.0
django-cors-headers==4.3.1
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-filter==23.3
drf-spectacular==0.26.5
django-extensions==3.2.3
whitenoise==6.6.0
Pillow==10.1.0
qrcode==7.4.2
